+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Unknown fields in the configuration will now cause an error instead of being ignored.
 - Loading a configuration file that does not exist will cause an error.
 - All of a signers sources are now queried concurrently.
+- If a user does not exist on a source a warning will be logged instead of returning an error.
 
 ## [0.4.1] - 2024-10-29